Get NPCs spawning and rendering. Move character animations into own directory. Bring over conversation, npc, and trainer files

+++ src/Core/Systems/Position.cpp
@@ -109,8 +109,8 @@ bl::entity::Entity& Position::get(const component::Position& p) {
         }
     }
 
-    BL_LOG_WARN << "Out of bounds entity check at (" << p.level << ", " << p.positionTiles().x
-                << ", " << p.positionTiles().y << ")";
+    BL_LOG_WARN << "Out of bounds entity check at (" << static_cast<int>(p.level) << ", "
+                << p.positionTiles().x << ", " << p.positionTiles().y << ")";
     null = bl::entity::InvalidEntity;
     return null;
 }
@@ -124,8 +124,8 @@ bl::entity::Entity Position::getEntity(const component::Position& p) const {
         }
     }
 
-    BL_LOG_WARN << "Out of bounds entity check at (" << p.level << ", " << p.positionTiles().x
-                << ", " << p.positionTiles().y << ")";
+    BL_LOG_WARN << "Out of bounds entity check at (" << static_cast<int>(p.level) << ", "
+                << p.positionTiles().x << ", " << p.positionTiles().y << ")";
     return bl::entity::InvalidEntity;
 }
